+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -24,8 +24,8 @@ import { default as freshserviceServer } from "@app/lib/api/actions/servers/fres
 import { default as frontServer } from "@app/lib/api/actions/servers/front";
 import { default as githubServer } from "@app/lib/api/actions/servers/github";
 import { default as gmailServer } from "@app/lib/api/actions/servers/gmail";
-import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
 import { default as gongServer } from "@app/lib/api/actions/servers/gong";
+import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
 import { default as driveServer } from "@app/lib/api/actions/servers/google_drive";
 import { default as sheetsServer } from "@app/lib/api/actions/servers/google_sheets";
 import { default as httpClientServer } from "@app/lib/api/actions/servers/http_client";
